requirements.txt
paho-mqtt==1.5.1
python-dotenv==0.15.0
python-ipmi==0.4.2
